requirements.txt
streamlit>=1.41.0
spacy>=3.8.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
openai>=1.0.0
